Minor fixes and Javadoc

+++ src/main/java/it/polimi/ingsw/server/model/leadercards/ZeroLeader.java
@@ -1,6 +1,5 @@
 package it.polimi.ingsw.server.model.leadercards;
 
-import it.polimi.ingsw.server.model.Production;
 import it.polimi.ingsw.server.model.cardrequirements.CardRequirement;
 import it.polimi.ingsw.server.model.resourcetypes.ResourceType;
 
@@ -43,12 +42,12 @@ public class ZeroLeader extends LeaderCard {
                     amountToConvert = Math.min(convertibleAmount, chosenAmount); // can't convert more than the lowest of the two
 
             /* Add converted resources */
-            resCopy.compute(this.getResource(), (res, amount) -> amount == null ? amountToConvert : amount.intValue() + amountToConvert);
+            resCopy.compute(this.getResource(), (res, amount) -> amount == null ? amountToConvert : amount + amountToConvert);
             
             /* Remove converted resources, deleting key if none left.
                If there's some left, zeros can be used in successive conversions; else it shouldn't be possible to do so,
                for that would transform more resources than it is allowed */
-            resCopy.compute(replaceableResType, (res, amount) -> amount == null || amount.intValue() - amountToConvert == 0 ? null : amount.intValue() - amountToConvert);
+            resCopy.compute(replaceableResType, (res, amount) -> amount == null || amount - amountToConvert == 0 ? null : amount - amountToConvert);
             replacements.compute(this.getResource(), (res, amount) -> amount - amountToConvert == 0 ? null : amount - amountToConvert);
         }
